# frontend/src/api/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "D&D Kids Resources API",
    "description": "API for D&D 5e resource management",
    "version": "2.0.0"
  },
  "paths": {
...
    "/api/dungeons/{dungeon_id}/session-state": {
      "get": {
        "tags": [
          "session_state"
        ],
        "summary": "Get Dungeon Session State",
        "description": "Get the permanent door/stair/portal toggle state for a dungeon",
        "operationId": "getDungeonSessionState",
        "parameters": [
          {
            "name": "dungeon_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Dungeon Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MapSessionStateBlob"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionStateError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "session_state"
        ],
        "summary": "Save Dungeon Session State",
        "description": "Save/upsert normalized runtime toggle state, or clear the row when nothing remains",
        "operationId": "saveDungeonSessionState",
        "parameters": [
          {
            "name": "dungeon_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Dungeon Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MapSessionStateBlob"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MapSessionStateBlob"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionStateError"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SessionStateError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "session_state"
        ],
        "summary": "Reset Dungeon Session State",
        "description": "Reset a dungeon's toggle state to its authored defaults (removes the saved row, if any)",
        "operationId": "resetDungeonSessionState",
        "parameters": [
          {
            "name": "dungeon_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Dungeon Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/loom/sessions": {
      "get": {
        "tags": [
          "loom"
        ],
        "summary": "List Sessions",
        "description": "List logged sessions in campaign order.",
        "operationId": "listSessions",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LoomSession"
                  },
                  "title": "Response Listsessions"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "loom"
        ],
        "summary": "Create Session",
        "description": "Create a session column on the tapestry.",
        "operationId": "createSession",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoomSessionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoomSession"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoomError"
                }
              }
            },
            "description": "Bad Request"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/loom/sessions/{session_id}": {
      "put": {
        "tags": [
          "loom"
        ],
        "summary": "Update Session",
        "description": "Update a session's title or date.",
        "operationId": "updateSession",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoomSessionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoomSession"
                }
              }
            }
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoomError"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoomError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "loom"
        ],
        "summary": "Delete Session",
        "description": "Delete a session; 422 while any node still belongs to it.",
        "operationId": "deleteSession",
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoomError"
                }
              }
            },
            "description": "Bad Request"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoomError"
                }
              }
            },
            "description": "Not Found"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    {
                      "$ref": "#/components/schemas/LoomError"
                    },
                    {
                      "$ref": "#/components/schemas/FastAPIValidationErrorResponse"
                    }
                  ],
                  "title": "Response 422 Deletesession"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "FastAPIValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Loc"
          },
          "msg": {
            "type": "string",
            "title": "Msg"
          },
          "type": {
            "type": "string",
            "title": "Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "FastAPIValidationError"
      },
      "FastAPIValidationErrorResponse": {
        "properties": {
          "detail": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/FastAPIValidationError"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "FastAPIValidationErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LoomError": {
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "session_ordinal_already_exists",
              "failed_to_create_session",
              "session_not_found",
              "failed_to_update_session",
              "cannot_delete_session_with_nodes",
              "failed_to_delete_session",
              "unknown_thread",
              "no_pending_beat",
              "invalid_session_log",
              "failed_to_log_session",
              "unknown_origin_node_id",
              "origin_node_must_reference_session",
              "thread_name_already_exists",
              "failed_to_create_thread",
              "thread_not_found",
              "failed_to_update_thread",
              "failed_to_delete_thread",
              "failed_to_create_node",
              "node_not_found",
              "node_kind_is_immutable",
              "failed_to_update_node",
              "cannot_delete_start_end_node",
              "failed_to_delete_node",
              "only_beats_can_be_fulfilled",
              "beat_must_be_placed_to_be_fulfilled",
              "beat_must_have_session_to_be_fulfilled",
              "only_beats_can_be_banked",
              "beat_is_not_placed_on_a_thread",
              "failed_to_bank_beat",
              "only_beats_or_sessions_can_be_placed",
              "node_already_in_thread",
              "node_already_on_another_thread",
              "failed_to_place_node",
              "node_not_a_member_of_thread",
              "start_end_position_is_fixed",
              "failed_to_reorder_node",
              "cannot_remove_start_end_nodes",
              "failed_to_remove_node_from_thread",
              "target_thread_not_found",
              "cannot_move_within_same_thread",
              "failed_to_move_node"
            ],
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "LoomError"
      },
...
      "LoomSession": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "ordinal": {
            "type": "integer",
            "title": "Ordinal"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "played_on": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Played On"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "id",
          "ordinal",
          "name"
        ],
        "title": "LoomSession"
      },
      "LoomSessionCreate": {
        "properties": {
          "ordinal": {
            "type": "integer",
            "title": "Ordinal"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "played_on": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Played On"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "ordinal",
          "name"
        ],
        "title": "LoomSessionCreate"
      },
...
      "LoomSessionUpdate": {
        "properties": {
          "ordinal": {
            "type": "integer",
            "title": "Ordinal"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "played_on": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Played On"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "ordinal",
          "name"
        ],
        "title": "LoomSessionUpdate"
      },
...
      "MapSessionStateBlob": {
        "properties": {
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "data"
        ],
        "title": "MapSessionStateBlob"
      },
...
      "SessionStateError": {
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "session_state_not_found",
              "dungeon_not_found",
              "failed_to_save_session_state"
            ],
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "SessionStateError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}